simplify working_dir default behavior
